requirements.txt
RPi.GPIO
pulsectl-asyncio
asyncio-mqtt